refactor: implement Phase 2 modular architecture with endpoint modules
Completed Phase 2 of the cBioPortal MCP server refactoring to create a
modular endpoint architecture with clean domain separation.

MAJOR ACHIEVEMENTS:
- Reduced main server from 1,352 to 371 lines (72.5% reduction\!)
- Exceeded planned target of 600-800 lines by achieving just 371 lines
- Created 4 focused endpoint modules with clear domain boundaries

NEW ENDPOINT MODULES:
- endpoints/studies.py - Cancer studies, types, search (~360 lines)
- endpoints/genes.py - Gene operations, mutations, batch processing (~240 lines)
- endpoints/samples.py - Sample data and sample lists (~80 lines)
- endpoints/molecular_profiles.py - Molecular profiles, clinical data (~280 lines)

ARCHITECTURE IMPROVEMENTS:
- Clean dependency injection pattern with APIClient
- All endpoint methods delegate to focused domain modules
- Maintained exact method signatures for FastMCP compatibility
- Clear separation of concerns across API domains

QUALITY ASSURANCE:
- All 92 tests continue to pass with zero regression
- Full API compatibility preserved across all endpoints
- Complete async performance benefits maintained (4.5x improvement)
- Comprehensive test coverage through careful delegation

Ready for Phase 3: Server class refactoring and further cleanup.
