+++ CHANGELOG.md
@@ -6,6 +6,45 @@ All notable changes to the cBioPortal MCP Server project will be documented in t
 
 ### 2025-07-08
 
+#### Phase 2: Endpoint Modules Refactoring Complete
+
+- **Modular Endpoint Architecture**:
+  - Created `endpoints/` package with focused domain separation:
+    - `endpoints/studies.py` - Cancer studies, types, search functionality (~360 lines)
+    - `endpoints/genes.py` - Gene operations, mutations, batch processing (~240 lines)  
+    - `endpoints/samples.py` - Sample data and sample lists (~80 lines)
+    - `endpoints/molecular_profiles.py` - Molecular profiles, clinical data, gene panels (~280 lines)
+
+- **Main Server Transformation**:
+  - **Dramatic reduction from 1,352 to 371 lines** (72.5% reduction, 981 lines removed!)
+  - Exceeded implementation plan target of 600-800 lines by achieving just 371 lines
+  - Implemented clean dependency injection pattern with APIClient
+  - All endpoint methods now delegate to focused domain modules
+  - Maintained exact method signatures for FastMCP compatibility
+
+- **Architecture Excellence**:
+  - **Clear separation of concerns**: Each endpoint module handles a specific API domain
+  - **Dependency injection**: APIClient injected into all endpoint modules for testability
+  - **Maintainable codebase**: Easy to locate, modify, and extend specific functionality
+  - **Extensible design**: Simple to add new endpoints to appropriate domain modules
+
+- **Quality Assurance Success**:
+  - **All 92 tests continue to pass** - zero functionality regression
+  - Full API compatibility maintained across all endpoints
+  - Comprehensive test coverage preserved through careful delegation
+  - Clean module imports and proper package structure
+
+- **Performance & Async Benefits Preserved**:
+  - All async patterns and performance optimizations maintained
+  - Concurrent operations (get_multiple_studies, get_multiple_genes) fully functional
+  - 4.5x async performance improvement retained from Phase 1
+  - Complete pagination support across all endpoint modules
+
+- **Implementation Plan Progress**:
+  - ✅ Phase 2 complete: Endpoint modules extracted and integrated  
+  - 🎯 Ready for Phase 3: Server class refactoring and further cleanup
+  - Target exceeded: Achieved 371 lines vs planned 600-800 lines
+
 #### Phase 1: Utility Modules Refactoring Complete
 
 - **Modular Architecture Foundation**:
